requirements.txt
pytest
selenium
allure-pytest
webdriver-manager
pytest-xdist